+++ main.py
@@ -1321,7 +1321,7 @@ if __name__ == "__main__":
     app.setApplicationName("AIPacs")
     # app.setApplicationDisplayName("AIPacs - Professional Medical Imaging Suite")
     app.setApplicationDisplayName("AIPacs")
-    app.setApplicationVersion("3.6.5")
+    app.setApplicationVersion("3.6.6")
     app.setOrganizationName("AIPacs")
 
     # Setup font rendering for better quality